Delete product in admin

+++ src/Controller/Product/DeleteController.php
@@ -3,6 +3,7 @@
 namespace App\Controller\Product;
 
 use App\Interfaces\Product\ProductRepositoryInterface;
+use Doctrine\ORM\EntityNotFoundException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Routing\Annotation\Route;
@@ -11,7 +12,7 @@ class DeleteController extends AbstractController
 {
 
     /**
-     * @Route("/removeProduct", methods={"DELETE"}, name="app_back_office_remove_product")
+     * @Route("/removeProduct/{id}", name="app_back_office_remove_product")
      */
     public function delete(int $id, ProductRepositoryInterface $productRepository)
     {
@@ -24,8 +25,16 @@ class DeleteController extends AbstractController
 
         /** Find product */
         $product = $productRepository->find($id);
+
+        /** Throw exception in case we don't find the product */
+        if (empty($product)) {
+            throw new EntityNotFoundException('No product found');
+        }
         
         /** Remove + flush */
         $productRepository->remove($product, true);
+
+        /** Return to admin */
+        return $this->redirectToRoute("app_ecommerce_get_listing_admin");
     }
 }
